+++ base/src/main/java/io/github/sds100/keymapper/base/constraints/ConstraintDependency.kt
@@ -11,7 +11,7 @@ enum class ConstraintDependency {
     WIFI_SSID,
     WIFI_STATE,
     CHOSEN_IME,
-    KEYBOARD_STATE,
+    KEYBOARD_VISIBLE,
     DEVICE_LOCKED_STATE,
     LOCK_SCREEN_SHOWING,
     PHONE_STATE,